docs: autodoc_mock_imports for pycurl
Adds mocking of `pycurl` that cannot be installed on ReadTheDocs
builders.  However, we would also have to remove `pycurl` as the
installation dependency of `cernopendata-client` in `setup.py`, which
is not wanted at this stage.

Hence amends CLI API page to simply instruct persons to use command
line to get the list of all commands and options. It is better to
show latest user guide instead.

This fixes the ReadTheDocs building problem.

Addresses #40.

Signed-off-by: Tibor Šimko <tibor.simko@cern.ch>